+++ CHANGELOG.md
@@ -13,8 +13,8 @@ v999 (February XX, 2021)
 
 **Developer changes**
 
-* Properly restrict access to statement history to users with system access, staff, admins.
-* Avoid name collissions when cloning a component.
+* Properly restrict statement history access to users with system, staff, or admin permissions.
+* Avoid name collisions when cloning a component.
 * Replaced function-based views with class-based listview for SelectedComponentsList, ProjectList.
 * Avoid name collisions when cloning a component.
 * Default to not use Django Debug Toolbar. Added new `enable_tool_bar` parameter option for `local/environment.json` to allow users to enable(True) or disable(False) the Django Debug Toolbar.